DTGB-907: Add the theme-body() themify function
The themify theme() function converts sass like this:

```scss
.parent {
  .child {
    theme('background-color', 'color-primary--lighten-4', 'field-background');
  }
}
```

gets transformed to css:

```css
.cs--blue .parent .child {
    background-color: #00f;
}
```

Where `.cs--blue` is the theme color class which is added to the
`<body>` element.

But sometimes the .parent selector is a body class which is not
supported by the `theme()` function.

The `theme-body()` function adds the color class to the first element of
the elector:

```scss
.parent {
  .child {
    theme-body('background-color', 'color-primary--lighten-4', 'field-background');
  }
}
```

gets transformed to css:

```css
.cs--blue.parent .child {
    background-color: #00f;
}
```